Add script to check empty field.
Add new script to control text on left part of start button to inform the facilitator if he forget to complete any field.
A swicth case control if the correct error message is wrote into the interface when function is called.
LoggingManager.cs script updated to call function when empty field is detected.