Fix int fields
Copy and adapt the implementation from the projet Whack A Mole, to simplify the code and understand in each project how the code work.
Now it is a button desactivate when you click on, and you can edit a int field, and the button display the value of the fied.