Change circle speed configuration
Change circle speed configuration.
Before, circle was destroy directly when your gaze hit 5 times the red circle, so it was necessary to reset circle time life to keep the original speed.
Now, circle speed is always the same one. To do this, just need to delete the reset of circle speedand condition to check if your gaze hit 5 times the red circle.